refactor: move server functions to server.ts
